backend/requirements.txt
blinker>=1.8.2
click>=8.1.7
Flask>=3.0.3
itsdangerous>=2.2.0
python-dotenv>=1.0.1
Werkzeug>=3.0.3
